python/requirements.txt
requests
beautifulsoup4
python-dotenv
instaloader
pandas
sentence-transformers
faiss-cpu
transformers
torch
scikit-learn
